+++ src/main/java/com/ctd/integrador/backend1/service/ITurnoService.java
@@ -6,9 +6,9 @@ import com.ctd.integrador.backend1.model.TurnoDTO;
 import java.util.Set;
 
 public interface ITurnoService {
-    TurnoDTO buscarPorId(Integer id) throws ResourceNotFoundException;
+    TurnoDTO buscarPorId(Long id) throws ResourceNotFoundException;
     Set<TurnoDTO> buscarTodos();
-    void eliminarTurno(Integer id) throws ResourceNotFoundException;
+    void eliminarTurno(Long id) throws ResourceNotFoundException;
     TurnoDTO actualizarTurno(TurnoDTO turnoDTO) throws ResourceNotFoundException;
     TurnoDTO agregarTurno(TurnoDTO turnoDTO);
 }
